Major refactoring, cleanup and optimization. WIP
