+++ prospector/git/git.py
@@ -15,16 +15,17 @@ from functools import lru_cache
 from urllib.parse import urlparse
 from dotenv import load_dotenv
 
+from git.exec import Exec
+from git.raw_commit import RawCommit
+
 import log.util
 
-# from pprint import pprint
-# import pickledb
 from stats.execution import execution_statistics, measure_execution_time
 
 _logger = log.util.init_local_logger()
 
 # If we don't parse .env file, we can't use the environment variables
-load_dotenv()
+# load_dotenv()
 
 GIT_CACHE = os.getenv("GIT_CACHE")
 
@@ -62,6 +63,7 @@ def clone_repo_multiple(
 def path_from_url(url, base_path):
     url = url.rstrip("/")
     parsed_url = urlparse(url)
+    print(parsed_url)
     return os.path.join(
         base_path, parsed_url.netloc + parsed_url.path.replace("/", "_")
     )
@@ -74,7 +76,7 @@ class Git:
         cache_path=os.path.abspath("/tmp/git-cache"),
         shallow=False,
     ):
-        self.repository_type = "GIT"
+        self._repository_type = "GIT"
         self._url = url
         self._path = path_from_url(url, cache_path)
         self._fingerprints = dict()
@@ -238,47 +240,39 @@ class Git:
             cmd = ["git", "rev-list"]
 
         if since:
-            cmd.append("--since=" + str(since))
+            cmd.append(f"--since={since}")
 
         if until:
-            cmd.append("--until=" + str(until))
+            cmd.append(f"--until={until}")
 
         if ancestors_of:
             cmd.append(ancestors_of)
 
         if exclude_ancestors_of:
-            cmd.append("^" + exclude_ancestors_of)
+            cmd.append(f"^{exclude_ancestors_of}")
 
         if filter_files:
-            cmd.append("*." + filter_files)
+            cmd.append(f"*.{filter_files}")
 
+        # What is this??
         if find_in_code:
-            cmd.append('-S"%s"' % find_in_code)
+            cmd.append(f"-S{find_in_code}")
 
         if find_in_msg:
-            cmd.append('--grep="%s"' % find_in_msg)
+            cmd.append(f"--grep={find_in_msg}")
 
         try:
             _logger.debug(" ".join(cmd))
             out = self._exec.run(cmd, cache=True)
-            #  cmd_test = ["git", "diff", "--name-only", self._id + "^.." + self._id]
-            # out = self._exec.run(cmd, cache=True)
+        #     --cherry-mark
+        # Like --cherry-pick (see below) but mark equivalent commits with =
+        # rather than omitting them, and inequivalent ones with +.
+
         except Exception:
             _logger.error("Git command failed, cannot get commits", exc_info=True)
             out = []
 
         out = [l.strip() for l in out]
-        # res = []
-        # try:
-        #     for id in out:
-        #         cmd = ["git", "diff", "--name-only", id + "^.." + id]
-        #         o = self._exec.run(cmd, cache=True)
-        #         for f in o:
-        #             if "mod_auth_digest" in f:
-        #                 res.append(id)
-        # except Exception:
-        #     _logger.error("Changed files retrieval failed", exc_info=True)
-        #     res = out
 
         return out
 
@@ -305,11 +299,11 @@ class Git:
     @measure_execution_time(execution_statistics.sub_collection("core"))
     def get_commit(self, key, by="id"):
         if by == "id":
-            return RawCommit(self, key)
+            return RawCommit(self._url, key, self._exec)
         if by == "fingerprint":
             # TODO implement computing fingerprints
             c_id = self._fingerprints[key]
-            return RawCommit(self, c_id)
+            return RawCommit(self._url, c_id, self._exec)
 
         return None
 
@@ -402,469 +396,6 @@ class Git:
         Return the text of the issue or PR with the given id
         """
         cmd = f"git fetch origin pull/{id}/head"
-
-
-class RawCommit:
-    def __init__(self, repository: Git, commit_id: str, init_data=None):
-        self._attributes = {}
-
-        self._repository = repository
-        self._id = commit_id
-        self._exec = repository._exec
-
-        # the following attributes will be initialized lazily and memoized, unless init_data is not None
-        if init_data:
-            for k in init_data:
-                self._attributes[k] = init_data[k]
-
-    def get_id(self) -> str:
-        if "full_id" not in self._attributes:
-            try:
-                cmd = ["git", "log", "--format=%H", "-n1", self._id]
-                self._attributes["full_id"] = self._exec.run(cmd, cache=True)[0]
-            except Exception:
-                _logger.error(
-                    f"Failed to obtain full commit id for: {self._id} in dir: {self._exec._workdir}",
-                    exc_info=True,
-                )
-        return self._attributes["full_id"]
-
-    def get_parent_id(self):
-        """
-        Returns the list of parents commits
-        """
-        if "parent_id" not in self._attributes:
-            try:
-                cmd = ["git", "log", "--format=%P", "-n1", self._id]
-                parent = self._exec.run(cmd, cache=True)[0]
-                parents = parent.split(" ")
-                self._attributes["parent_id"] = parents
-            except:
-                _logger.error(
-                    f"Failed to obtain parent id for: {self._id} in dir: {self._exec._workdir}",
-                    exc_info=True,
-                )
-        return self._attributes["parent_id"]
-
-    def get_repository(self):
-        return self._repository._url
-
-    def get_msg(self):
-        if "msg" not in self._attributes:
-            self._attributes["msg"] = ""
-            try:
-                cmd = ["git", "log", "--format=%B", "-n1", self._id]
-                self._attributes["msg"] = " ".join(self._exec.run(cmd, cache=True))
-            except Exception:
-                _logger.error(
-                    f"Failed to obtain commit message for commit: {self._id} in dir: {self._exec._workdir}",
-                    exc_info=True,
-                )
-        return self._attributes["msg"]
-
-    def get_diff(self, context_size: int = 1, filter_files: str = ""):
-        if "diff" not in self._attributes:
-            self._attributes["diff"] = ""
-            try:
-                cmd = [
-                    "git",
-                    "diff",
-                    "--unified=" + str(context_size),
-                    self._id + "^.." + self._id,
-                ]
-                if filter_files:
-                    cmd.append("*." + filter_files)
-                self._attributes["diff"] = self._exec.run(cmd, cache=True)
-            except Exception:
-                _logger.error(
-                    f"Failed to obtain patch for commit: {self._id} in dir: {self._exec._workdir}",
-                    exc_info=True,
-                )
-        return self._attributes["diff"]
-
-    def get_timestamp(self, date_format=None):
-        if "timestamp" not in self._attributes:
-            self._attributes["timestamp"] = None
-            self.get_timing_data()
-            # self._timestamp = self.timing_data()[2]
-        if date_format:
-            return datetime.utcfromtimestamp(
-                int(self._attributes["timestamp"])
-            ).strftime(date_format)
-        return int(self._attributes["timestamp"])
-
-    @measure_execution_time(
-        execution_statistics.sub_collection("core"),
-        name="retrieve changed file from git",
-    )
-    def get_changed_files(self):
-        if "changed_files" not in self._attributes:
-            cmd = ["git", "diff", "--name-only", self._id + "^.." + self._id]
-            try:
-                out = self._exec.run(cmd, cache=True)
-                self._attributes["changed_files"] = out  # This is a tuple
-            # This exception is raised when the commit is the first commit in the repository
-            except Exception as e:
-                _logger.error(
-                    f"Failed to obtain changed files for commit {self._id}, it may be the first commit of the repository. Processing anyway...",
-                    exc_info=True,
-                )
-                self._attributes["changed_files"] = []
-
-        return self._attributes["changed_files"]
-
-    def get_changed_paths(self, other_commit=None, match=None):
-        # TODO refactor, this overlaps with changed_files
-        if other_commit is None:
-            other_commit_id = self._id + "^"
-        else:
-            other_commit_id = other_commit._id
-
-        cmd = [
-            "git",
-            "log",
-            "--name-only",
-            "--format=%n",
-            "--full-index",
-            other_commit_id + ".." + self._id,
-        ]
-        try:
-            out = self._exec.run(cmd, cache=True)
-        except Exception as e:
-            out = str()
-            sys.stderr.write(str(e))
-            sys.stderr.write(
-                "There was a problem when getting the list of commits in the interval %s..%s\n"
-                % (other_commit.id()[0], self._id)
-            )
-            return out
-
-        if match:
-            out = [l.strip() for l in out if re.match(match, l)]
-        else:
-            out = [l.strip() for l in out]
-
-        return out
-
-    def get_hunks(self, grouped=False):
-        def is_hunk_line(line):
-            return line[0] in "-+" and (len(line) < 2 or (line[1] != line[0]))
-
-        def flatten_groups(hunk_groups):
-            hunks = []
-            for group in hunk_groups:
-                for h in group:
-                    hunks.append(h)
-            return hunks
-
-        def is_new_file(l):
-            return l[0:11] == "diff --git "
-
-        if "hunks" not in self._attributes:
-            self._attributes["hunks"] = []
-
-            diff_lines = self.get_diff()
-            # pprint(diff_lines)
-
-            first_line_of_current_hunk = -1
-            current_group = []
-            line_no = 0
-            for line_no, line in enumerate(diff_lines):
-                # print(line_no, line)
-                if is_new_file(line):
-                    if len(current_group) > 0:
-                        self._attributes["hunks"].append(current_group)
-                        current_group = []
-                        first_line_of_current_hunk = -1
-
-                elif is_hunk_line(line):
-                    if first_line_of_current_hunk == -1:
-                        # print('first_line_of_current_hunk', line_no)
-                        first_line_of_current_hunk = line_no
-                else:
-                    if first_line_of_current_hunk != -1:
-                        current_group.append((first_line_of_current_hunk, line_no))
-                        first_line_of_current_hunk = -1
-
-            if first_line_of_current_hunk != -1:
-                # wrap up hunk that ends at the end of the patch
-                # print('line_no:', line_no)
-                current_group.append((first_line_of_current_hunk, line_no + 1))
-
-            self._attributes["hunks"].append(current_group)
-
-        if grouped:
-            return self._attributes["hunks"]
-        else:
-            return flatten_groups(self._attributes["hunks"])
-
-    def equals(self, other_commit):
-        """
-        Return true if the two commits contain the same changes (despite different commit messages)
-        """
-        return self.get_fingerprint() == other_commit.get_fingerprint()
-
-    def get_fingerprint(self):
-        if "fingerprint" not in self._attributes:
-            # try:
-            cmd = ["git", "show", '--format="%t"', "--numstat", self._id]
-            out = self._exec.run(cmd, cache=True)
-            self._attributes["fingerprint"] = hashlib.md5(
-                "\n".join(out).encode()
-            ).hexdigest()
-
-        return self._attributes["fingerprint"]
-
-    def get_timing_data(self):
-        data = self._get_timing_data()
-        self._attributes["next_tag"] = data[0]
-        # self._next_tag = data[0]
-
-        self._attributes["next_tag_timestamp"] = data[1]
-        # self._next_tag_timestamp = data[1]
-
-        self._attributes["timestamp"] = data[2]
-        # self._timestamp = data[2]
-
-        self._attributes["time_to_tag"] = data[3]
-        # self._time_to_tag = data[3]
-
-    # TODO refactor
-    # this method should become private and should be invoked to initialize (lazily)
-    # # the relevant attributes.
-    def _get_timing_data(self):
-        # print("WARNING: deprecated method Commit::timing_data(), use Commit::get_next_tag() instead.")
-        # if not os.path.exists(self._path):
-        #     print('Folder ' + self._path + ' must exist!')
-        #     return None
-
-        # get tag info
-        raw_out = self._exec.run(
-            "git tag --sort=taggerdate --contains " + self._id, cache=True
-        )  # ,  cwd=self._path)
-        if raw_out:
-            tag = raw_out[0]
-            tag_timestamp = self._exec.run(
-                'git show -s --format="%at" ' + tag + "^{commit}", cache=True
-            )[0][1:-1]
-        else:
-            tag = ""
-            tag_timestamp = "0"
-
-        try:
-            commit_timestamp = self._exec.run(
-                'git show -s --format="%at" ' + self._id, cache=True
-            )[0][1:-1]
-            time_delta = int(tag_timestamp) - int(commit_timestamp)
-            if time_delta < 0:
-                time_delta = -1
-        except:
-            commit_timestamp = "0"
-            time_delta = 0
-
-        # tag_date = datetime.utcfromtimestamp(int(tag_timestamp)).strftime(
-        #     "%Y-%m-%d %H:%M:%S"
-        # )
-        # commit_date = datetime.utcfromtimestamp(int(commit_timestamp)).strftime(
-        #     "%Y-%m-%d %H:%M:%S"
-        # )
-
-        # if self._verbose:
-        #     print("repository:                 " + self._repository._url)
-        #     print("commit:                     " + self._id)
-        #     print("commit_date:                " + commit_timestamp)
-        #     print("                            " + commit_date)
-        #     print("tag:                        " + tag)
-        #     print("tag_timestamp:              " + tag_timestamp)
-        #     print("                            " + tag_date)
-        #     print(
-        #         "Commit-to-release interval: {0:.2f} days".format(
-        #             time_delta / (3600 * 24)
-        #         )
-        #     )
-
-        self._timestamp = commit_timestamp
-        return (tag, tag_timestamp, commit_timestamp, time_delta)
-
-    def get_tags(self):
-        if "tags" not in self._attributes:
-            cmd = "git tag --contains " + self._id
-            tags = self._exec.run(cmd, cache=True)
-            if not tags:
-                self._attributes["tags"] = []
-            else:
-                self._attributes["tags"] = tags
-
-        return self._attributes["tags"]
-
-    def get_next_tag(self):
-        if "next_tag" not in self._attributes:
-            self.get_timing_data()
-        return (
-            self._attributes["next_tag"],
-            self._attributes["next_tag_timestamp"],
-            self._attributes["time_to_tag"],
-        )
-
-    def __str__(self):
-        data = (
-            self._id,
-            self.get_timestamp(date_format="%Y-%m-%d %H:%M:%S"),
-            self.get_timestamp(),
-            self._repository.get_url(),
-            self.get_msg(),
-            len(self.get_hunks()),
-            len(self.get_changed_paths()),
-            self.get_next_tag()[0],
-            "\n".join(self.get_changed_paths()),
-        )
-        return """
-        Commit id:         {}
-        Date (timestamp):  {} ({})
-        Repository:        {}
-        Message:           {}
-        hunks: {},  changed files: {},  (oldest) tag: {}
-        {}""".format(
-            *data
-        )
-
-
-class RawCommitSet:
-    def __init__(self, repo=None, commit_ids=[], prefetch=False):
-
-        if repo is not None:
-            self._repository = repo
-        else:
-            raise ValueError  # pragma: no cover
-
-        self._commits = []
-
-        # TODO when the flag 'prefetch' is True, fetch all data in one shot (one single
-        # call to the git binary) and populate all commit objects. A dictionary paramenter
-        # passed to the Commit constructor will be used to pass the fields that need to be populated
-        commits_count = len(commit_ids)
-        if prefetch is True and commits_count > 50:
-            _logger.warning(
-                f"Processing {commits_count:d} commits will take some time!"
-            )
-            for cid in commit_ids:
-                commit_data = {"id": "", "msg": "", "patch": "", "timestamp": ""}
-                current_field = None
-                commit_raw_data = self._repository._exec.run(
-                    "git show --format=@@@@@SHA1@@@@@%n%H%n@@@@@LOGMSG@@@@@%n%s%n%b%n@@@@@TIMESTAMP@@@@@@%n%at%n@@@@@PATCH@@@@@ "
-                    + cid,
-                    cache=True,
-                )
-
-                for line in commit_raw_data:
-                    if line == "@@@@@SHA1@@@@@":
-                        current_field = "id"
-                    elif line == "@@@@@LOGMSG@@@@@":
-                        current_field = "msg"
-                    elif line == "@@@@@TIMESTAMP@@@@@":
-                        current_field = "timestamp"
-                    else:
-                        commit_data[current_field] += "\n" + line
-
-                self._commits.append(
-                    RawCommit(self._repository, cid, init_data=commit_data)
-                )
-        else:
-            self._commits = [RawCommit(self._repository, c) for c in commit_ids]
-
-    def get_all(self):
-        return self._commits
-
-    def add(self, commit_id):
-        self._commits.append(RawCommit(self._repository, commit_id))
-        return self
-
-    def filter_by_msg(self, word):
-        return [c for c in self.get_all() if word in c.get_msg()]
-
-
-class Exec:
-    def __init__(self, workdir=None, encoding="latin-1", timeout=None):
-        self._encoding = encoding
-        self._timeout = timeout
-        self.setDir(workdir)
-
-    def setDir(self, path):
-        if os.path.isabs(path):
-            self._workdir = path
-        else:
-            raise ValueError("Path must be absolute for Exec to work: " + path)
-
-    def run(self, cmd, ignore_output=False, cache: bool = False):
-        if cache:
-            result = self._run_cached(
-                tuple(cmd) if isinstance(cmd, list) else cmd, ignore_output
-            )
-        else:
-            result = self._run_uncached(
-                tuple(cmd) if isinstance(cmd, list) else cmd, ignore_output
-            )
-        return result
-
-    @lru_cache(maxsize=10000)
-    def _run_cached(self, cmd, ignore_output=False):
-        return self._run_uncached(cmd, ignore_output=ignore_output)
-
-    def _run_uncached(self, cmd, ignore_output=False):
-        if isinstance(cmd, str):
-            cmd = cmd.split()
-
-        if ignore_output:
-            self._execute_no_output(cmd)
-            return ()
-
-        result = self._execute(cmd)
-        if result is None:
-            return ()
-
-        return tuple(result)
-
-    def _execute_no_output(self, cmd_l):
-        try:
-            subprocess.check_call(
-                cmd_l,
-                cwd=self._workdir,
-                stdout=subprocess.DEVNULL,
-                stderr=subprocess.DEVNULL,
-            )
-        except subprocess.TimeoutExpired:  # pragma: no cover
-            _logger.error(
-                "Timeout exceeded (" + self._timeout + " seconds)", exc_info=True
-            )
-            raise Exception("Process did not respond for " + self._timeout + " seconds")
-
-    def _execute(self, cmd_l):
-        try:
-            proc = subprocess.Popen(
-                cmd_l,
-                cwd=self._workdir,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.STDOUT,  # Needed to have properly prinded error output
-            )
-            out, _ = proc.communicate()
-
-            if proc.returncode != 0:
-                raise Exception(
-                    "Process (%s) exited with non-zero return code" % " ".join(cmd_l)
-                )
-            # if err:       # pragma: no cover
-            #     traceback.print_exc()
-            #     raise Exception('Execution failed')
-
-            raw_output_list = out.decode(self._encoding).split("\n")
-            return [r for r in raw_output_list if r.strip() != ""]
-        except subprocess.TimeoutExpired:  # pragma: no cover
-            _logger.error(f"Timeout exceeded ({self._timeout} seconds)", exc_info=True)
-            raise Exception(f"Process did not respond for {self._timeout} seconds")
-            # return None
-        # except Exception as ex:                 # pragma: no cover
-        #     traceback.print_exc()
-        #     raise ex
 
 
 # Donald Knuth's "reservoir sampling"
