+++ prospector/git/version_to_tag.py
@@ -10,7 +10,8 @@ import difflib
 # pylint: disable=singleton-comparison,unidiomatic-typecheck, dangerous-default-value
 import re
 
-from .git import RawCommit, Git
+from git.raw_commit import RawCommit
+from git.git import Git
 
 
 def recursively_split_version_string(input_version: str, output_version: list = []):
